lxd/backup: Move config struct into own package to use in non-cgo applications
Otherwise the import of the db package prevents use of it in non-cgo applications.

Moves the ToInstanceDBArgs function into backup.ConfigToInstanceDBArgs().
